refactor: PrezelAvatar 내부 로직 및 스타일 함수 수정
*   `prezelAvatarContainerSize`, `prezelAvatarIconSize` 함수에서 불필요한 `@Composable` 어노테이션을 제거했습니다.
*   `PrezelAvatar` 내부의 `isError` 상태 위치를 `PrezelAvatarType.Image` 분기 내부로 이동하고, `url` 변경 시 상태가 초기화되도록 `remember(type.url)`을 적용했습니다.
*   미리보기(Preview) 데이터의 URL 프로토콜을 `http`에서 `https`로 수정했습니다.
